Client side API beginning to take shape
The client-side API, as of now, is made up of two primary classes:
"TM_Client" and "TM_Share".

TM_Share represents the local form of a loaded piece of shared
transactional memory. It will eventually include many overloaded
operators for easy use. Currently, it only includes the assignment
operator. These overloads have a much more important role than
abstracting standard operations. When these operations are made, the
TM_Share class must log whether it was a read or a write. This
information is logged to a TM_Message struct which is pushed to a
referenced queue. This queue will be watched concurrently by the thread
dedicated to keeping the TM_Host informed of the client's actions.

TM_Client is the primary class, the class most directly interacted with
by a developer. It will likely become a mostly static class in the
future, as separate instances of it in a single program are beyond the
scope and need of this project. This class includes the "Transaction"
struct, which serves as a container and holds an abort count, name, and
a function pointer. The function pointer is assigned to a users function
and will be executed when a transaction is called. The user registers a
transaction and then executes it using either it's id or name. This
TM_Client class also holds the vector of these Transaction structs and
the queue of TM_Messages that will be referenced by the TM_Shares.
Additionally, a vector or TM_Shares will be included in the Transaction
struct.

The user will define functions as "void* foo(void*)" for use with the
TM_Client. This is very similar to the pthread API. This function can be
wrapped in an exception handler using #define macros, allowing the
TM_Share and the TM_Client to throw exceptions to end the functions
execution prematurely. This will effectively implement the eager
conflict management.

The client API is still very early in its development, many features are
still required, as well as testing.