Exception methodology tested
The main.cpp now performs a simple test of the current implementation,
the test proves that the #define keywords and exception handling work as
expected.

Communication with the host will be difficult to do asynchronously while
maintaining eager conflict management. There are two very feasible
options, demonstrated by the addition of the auto_sync private member.

[Note that host synchronization is not implemented in this commit; auto_sync does
nothing at this point]

(1) With auto_sync on, every write to shared memory will result in a
call to the corresponding network communication back-end to inform the
host that a write has occurred. This will make transactional writes very
costly as the code must wait on a response from the host.

(2) User defined sync points. The developer can place in their code
calls to sync and check that their is not a conflict. This could be
simplified with #definec.

Still to be considered is how the host will abort a transaction that is
not currently syncing. If it is not possible it will only hurt
performance and priority possibilities, not overall functionality.