TM_Client is now static
The TM_Client is now static and allows our API to easily use its
interfaces without requiring the developer to create an instance of it.
This works very well for the macros that are being used. The BEGIN_T
macro can now transparently get a vector of TM_Share for the developer
to work on.