Created Transaction Class
Transaction is now a class, instead of a struct. This is primarily
because we may want to add convienent getters and setters here as well
as potentially store original state of shared memory while speculative
changes are made. This would allow the host force a restart and only
send the memory states that may be needed.