Refactoring; Networking tests implemented
On write operations, the TM_Share will now send a "hi!" message to the
server, for testing. On deconstruction, the TM_Client class will send
"SHUTDOWN" to indicate to the server thread that it is terminating.

The main file facilitates these test by proceeding only after user input,
making the tests easier to do.

The functionality listed is working as of this commit.

Debug flag was added to the makefile compilation to add gdb symbols.