SendMessage() sends real data
The SendMessage() method in TM_Share now sends all the TM_Message
information of the network. It uses sprintf to create a formatted string
with the members of the message. The the different message values are
separated by colons

<CODE>:<DATA>:<ADDRESS>

Some the type of "data" was changed to be an unsigned integer instead of
a character pointer.