TM_Share::ReceiveMessage()
Receives and parses a formatted message from server into a TM_Message
and returns it. Not fully implemented. Several other variables were
moved to private variables in TM_Share.

out_buffer is a char array because it is simple to use sprintf() to
create a format c-string (character array). in_buffer is instead a
string because it is easier to piece apart a C++ string using its public
methods. At least that's how it is for now...