Stream overload; overload bugs; TM_Share methods
Add a stream overloadd to the TM_Share, which prints either the
speculative value (if isWrite is set, meaning it was allowed to write)
or the original (old) value.

The other overloaded operators weren't using these values correctly,
corrections were made for these overloads (assign to new_value on writes
and set the isWrite member to true).

Added TM_Share::Sync(), which will send the sync code OR'd with the
current state (read or write) to the server. This indicates to the
server that the client wants it to check for any conflicts that may have
arisen since the operation and take the necessary action (abort or not).
No value need to be sent, as this is NOT a commit.

Added other prototypes to TM_Share: Init() and Commit(). These are self
explanatory but unimplemented at this point.