INIT Code; Transaction::Sync() added
An INIT code has been added to the protocol for initialization of the
memory; actually loading the current value of the address in the TM.

Transaction::Sync() simply iterates through its shared memory calling
Sync() on each. This is wrapped in a try/catch to handle any aborts due
to the sync.