Name announce; Commit; Init; Support methods
Client now indicates its name to the server upon connection. The commit
implementation is two phase, first the transaction uses static methods
in TM_Share to announce to the server its attempt to commit the
transaction. If that announcement does not get denied, then the commit
will be successful and the client can begin sending data. This is an
attempt to try and limit the data movement over the network.

TM_Init, which is called at the begining of a transaction, fetches and
intializes the share memory structures.