Commit now handles READ finishes too
When commiting a shared value that has only been read, the TM_Share
class will notify the server. The server will simply clear the read set
for that processor on that address. This prevents the need to store or
search for read sets after a commit (which would also require us to save
commit state).