Send NULL terminator; debug statements
The size of the string returned by sprintf does not include the NULL
terminator, therefore, the byte size needs one added to it to make sure
that the NULL terminator is also send through to the server. This
problem was fixed in several places. It's is still unclear whether
this was the cause of some bugs.

Also, severl debug statements were added.